fix: add guard for deletion depending on if CR owns apiservice

+++ controllers/gitopspromoter/apiservice.go
@@ -56,10 +56,17 @@ func ReconcilePromoterAPIServerAPIService(client client.Client, compName string,
 	}
 
 	if exists {
+		// Check to see if referenced service was created by the CR if it is not do not do any operations on it
+		// This safe guards from another ArgoCD CR instance from deleting the APIService even if the promoter is disabled
+		// If the service field is for some reason nil be safe and skip deleting anyway
+		if !ownsAPISvc(compName, cr, apiSvc) {
+			return apiSvc, nil
+		}
+
 		if !cr.Spec.Promoter.IsEnabled() || !enabled || !allowed {
 			argoutil.LogResourceDeletion(log, apiSvc, fmt.Sprintf("promoter apiservice for component %s is being deleted due to being disabled", compName))
 			if err := client.Delete(context.Background(), apiSvc); err != nil {
-				return nil, fmt.Errorf("failed to delete promoter service %s: %v", apiSvc.Name, err)
+				return nil, fmt.Errorf("failed to delete promoter service api %s: %v", apiSvc.Name, err)
 			}
 			return apiSvc, nil
 		}
@@ -149,12 +156,20 @@ func buildAPIServiceSpec(client client.Client, compName string, cr *argoproj.Arg
 }
 
 // DeleteAPIServices deletes a list of API Services
-func DeleteAPIServices(c client.Client, apiSvcList *apiregistrationv1.APIServiceList) error {
+func DeleteAPIServices(c client.Client, apiSvcList *apiregistrationv1.APIServiceList, compName string, cr *argoproj.ArgoCD) error {
 	for _, apiSvc := range apiSvcList.Items {
+		if !ownsAPISvc(compName, cr, &apiSvc) {
+			continue
+		}
+
 		argoutil.LogResourceDeletion(log, &apiSvc, "cleaning up cluster resources")
 		if err := c.Delete(context.TODO(), &apiSvc); err != nil {
 			return fmt.Errorf("failed to delete APIService %s during cleanup: %w", apiSvc.Name, err)
 		}
 	}
 	return nil
+}
+
+func ownsAPISvc(compName string, cr *argoproj.ArgoCD, apiSvc *apiregistrationv1.APIService) bool {
+	return apiSvc.Spec.Service != nil && cr.Namespace == apiSvc.Spec.Service.Namespace && generatePromoterResourceName(compName, cr) == apiSvc.Spec.Service.Name
 }
